Refactor: Components(=PhotoLetter) 분리 작업
